codegen: only emit calls to on_instructions when debugging
We now only emit calls to `Runtime::on_instruction` when GDB is
attached, or instruction tracing is enabled. Interrupts are instead
polled at the beginning of a block, where the bus gets ticked as well.

This provides a significant speedup, as we do not need to trap into the
runtime environment nearly as much.