jit: update inkwell, make everything fallible
This allows us to switch to LLVM 17 in the future, as support for that
has been implemented in the latest revision of inkwell.

It also makes practically every JIT code generation fallible, as a
side-effect of inkwell having done the same thing for all the `build_x`
methods. This should result in more robust error handling.